requirements.txt
# Биржа
ccxt==4.2.15

# ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
joblib==1.3.2

# HuggingFace
huggingface-hub==0.21.4

# Google Sheets
gspread==6.0.2
google-auth==2.28.2

# Telegram
python-telegram-bot==20.8

# Планировщик
schedule==1.2.1

# HTTP запросы
requests==2.31.0

# Технический анализ
ta==0.11.0

# Веб-сервер (для Render healthcheck)
flask==3.0.2